+++ packages/shared/src/lib/weightConversion.js
@@ -1,11 +1,13 @@
+import { parseDecimal } from './numberUtils.js'
+
 const LB_TO_KG = 0.45359237
 const KG_TO_LB = 2.20462262
 
 export function convertWeight(value, mode) {
-  if (value === '' || isNaN(parseFloat(value))) return null
-  const num = parseFloat(value)
+  if (value === '' || isNaN(parseDecimal(value))) return null
+  const num = parseDecimal(value)
   const result = mode === 'lb-to-kg' ? num * LB_TO_KG : num * KG_TO_LB
-  return parseFloat(result.toFixed(2))
+  return parseDecimal(result.toFixed(2))
 }
 
 export function getWeightUnits(mode) {
